+++ R/utils.R
@@ -8,6 +8,7 @@ select_loc <- function(x,
                        allow_rename = TRUE,
                        allow_empty = TRUE,
                        allow_predicates = TRUE,
+                       error_arg = NULL,
                        error_call = current_env()) {
   check_dots_empty()
 
@@ -21,6 +22,7 @@ select_loc <- function(x,
     allow_rename = allow_rename,
     allow_empty = allow_empty,
     allow_predicates = allow_predicates,
+    error_arg = error_arg,
     error_call = error_call
   )
 }